# [AI Accelerators] softmax kernel for Nested Tensor (CPU)